+++ src/config/apollo-client/apolloClient.ts
@@ -5,20 +5,24 @@ import { getMainDefinition } from '@apollo/client/utilities'
 import { createClient } from 'graphql-ws'
 
 import { __development__ } from '../../constants'
+import { toInt } from '../../utils'
 import { getAppEndPoint } from '../domain'
 import { cache } from './apollo-client-cache'
 
 const retryLink = new RetryLink({
   attempts(count, _, error) {
     const err = error?.result?.error
+    const statusCode = toInt(`${error?.statusCode}`)
+
     return (
-      err &&
-      Boolean(
-        (err.code === 'STALE_REFRESH_TOKEN' ||
-          err.code === 'EXPIRED_REFRESH_TOKEN' ||
-          err.code === 'INVALIDE_REFRESH_TOKEN') &&
-          count <= 2,
-      )
+      count <= 2 &&
+      (Boolean(
+        err &&
+          (err.code === 'STALE_REFRESH_TOKEN' ||
+            err.code === 'EXPIRED_REFRESH_TOKEN' ||
+            err.code === 'INVALIDE_REFRESH_TOKEN'),
+      ) ||
+        statusCode === 408)
     )
   },
   delay: {
@@ -27,7 +31,6 @@ const retryLink = new RetryLink({
 })
 
 const apiServiceEndPoint = getAppEndPoint()
-
 const httpLink = createHttpLink({
   uri: `${apiServiceEndPoint}/graphql`,
   credentials: 'include',