+++ src/graphql/fragments/project.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client'
 
 import { FRAGMENT_ENTRY_FOR_PROJECT } from './entries'
-import { FRAGMENT_USER_FOR_AVATAR, FRAGMENT_USER_ME } from './user'
+import { FRAGMENT_PROJECT_OWNER_USER, FRAGMENT_USER_FOR_AVATAR, FRAGMENT_USER_ME } from './user'
 
 export const FRAGMENT_PROJECT_NOSTR_KEYS = gql`
   fragment ProjectNostrKeys on Project {
@@ -105,13 +105,15 @@ export const FRAGMENT_PROJECT_REWARD_FOR_CREATE_UPDATE = gql`
     isHidden
     category
     preOrder
+    estimatedAvailabilityDate
+    estimatedDeliveryInWeeks
   }
 `
 
 export const FRAGMENT_PROJECT = gql`
   ${FRAGMENT_PROJECT_REWARD_FOR_CREATE_UPDATE}
   ${FRAGMENT_ENTRY_FOR_PROJECT}
-  ${FRAGMENT_USER_ME}
+  ${FRAGMENT_PROJECT_OWNER_USER}
   ${FRAGMENT_USER_FOR_AVATAR}
   fragment Project on Project {
     id
@@ -144,7 +146,7 @@ export const FRAGMENT_PROJECT = gql`
     owners {
       id
       user {
-        ...UserMe
+        ...ProjectOwnerUser
       }
     }
     rewards {