+++ src/styles/colors.ts
@@ -94,7 +94,7 @@ export const secondaryColors = {
   green: '#00E34D',
   red: '#DF3634',
   yellow: '#FFE600',
-  orange: '#FB923C',
+  orange: '#DD6B20',
   pink: '#FC257A',
   blue: '#438FF2',
   purple: '#670C8B',
@@ -138,5 +138,3 @@ export const darkModeColors = {
   social: socialColorsDark,
   app: ColorsAppDark,
 }
-
-export const colorOrange = '#DD6B20'